ConnectionForm now clears its own form tracking Add hook to generate unique form id Update form tracker component to use new hook if no id is specified
